Week3
Fell behind on using git. Completed the 3rd week of lectures and
assignment 3. I couldn’t figure out how to do assignment 2. There
seemed to be a general consensus
on the forums that there was something going on with github that was
not allowing us to download the repository. I will try again after I
finish the 4th week.